rename xgboost.trees module to tabular_trees.xgboost
